Added DefaultTypeDelegate
This component will be used by RemoteProxy when a service goes down. RemoteProxy passes an existing cached key of data into this component, and the component examines the data type, in an attempt to return adequate default results to the client, in the absence of the service.

Initial rev. supports: Query, Struct